Move ProductVariant to Variant
In Solidus variants model is named `Variant`; let’s keep the GraphQL
type consistent.
